Update README.md
The goal of this project is to teach students the necessary analytical skills to refactor existing open-source Java projects. Students will choose an open-source Java project, either from GitHub or one they have developed in previous coursework, and apply Java creational, structural, and behavioural design patterns to enhance the project's maintainability, extensibility, and overall code quality. Students must choose a Java open-source project of an acceptable size that is suitable for refactoring. The open-source project has to include a minimum of 5 unique classes that are not tied together in an inheritance relationship. The project deliverables are organised into four stages. In the following, the project deliverables for each stage are described with their respective deadlines